Refactoring processors
