+++ src/main/java/org/ballerinax/kubernetes/KubernetesAnnotationProcessor.java
@@ -19,7 +19,6 @@
 package org.ballerinax.kubernetes;
 
 import org.apache.commons.codec.binary.Base64;
-import org.ballerinalang.model.tree.AnnotationAttachmentNode;
 import org.ballerinax.kubernetes.exceptions.KubernetesPluginException;
 import org.ballerinax.kubernetes.handlers.ConfigMapHandler;
 import org.ballerinax.kubernetes.handlers.DeploymentHandler;
@@ -39,16 +38,11 @@ import org.ballerinax.kubernetes.models.PodAutoscalerModel;
 import org.ballerinax.kubernetes.models.SecretModel;
 import org.ballerinax.kubernetes.models.ServiceModel;
 import org.ballerinax.kubernetes.utils.KubernetesUtils;
-import org.wso2.ballerinalang.compiler.tree.BLangAnnotationAttachment;
-import org.wso2.ballerinalang.compiler.tree.expressions.BLangArrayLiteral;
-import org.wso2.ballerinalang.compiler.tree.expressions.BLangExpression;
-import org.wso2.ballerinalang.compiler.tree.expressions.BLangLiteral;
 import org.wso2.ballerinalang.compiler.tree.expressions.BLangRecordLiteral;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -61,7 +55,8 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 
-import static org.ballerinax.kubernetes.utils.KubernetesUtils.resolveValue;
+import static org.ballerinax.kubernetes.KubernetesConstants.INGRESS_FILE_POSTFIX;
+import static org.ballerinax.kubernetes.utils.KubernetesUtils.isEmpty;
 
 /**
  * Process Kubernetes Annotations and generate Artifacts.
@@ -71,36 +66,16 @@ class KubernetesAnnotationProcessor {
     private static final String DOCKER = "docker";
     private static final String BALX = ".balx";
     private static final String DEPLOYMENT_POSTFIX = "-deployment";
-    private static final String SVC_POSTFIX = "-svc";
-    private static final String INGRESS_POSTFIX = "-ingress";
     private static final String HPA_POSTFIX = "-hpa";
     private static final String DEPLOYMENT_FILE_POSTFIX = "_deployment";
     private static final String SVC_FILE_POSTFIX = "_svc";
     private static final String SECRET_FILE_POSTFIX = "_secret";
     private static final String CONFIG_MAP_FILE_POSTFIX = "_config_map";
     private static final String VOLUME_CLAIM_FILE_POSTFIX = "_volume_claim";
-    private static final String INGRESS_FILE_POSTFIX = "_ingress";
     private static final String HPA_FILE_POSTFIX = "_hpa";
     private static final String YAML = ".yaml";
     private static final String DOCKER_LATEST_TAG = ":latest";
-    private static final String INGRESS_HOSTNAME_POSTFIX = ".com";
     private PrintStream out = System.out;
-
-    /**
-     * Generate map by splitting keyValues.
-     *
-     * @param keyValues key value paris.
-     * @return Map of key values.
-     */
-    private Map<String, String> getMap(List<BLangRecordLiteral.BLangRecordKeyValue> keyValues) {
-        Map<String, String> map = new HashMap<>();
-        if (keyValues != null) {
-            keyValues.forEach(keyValue -> {
-                map.put(keyValue.getKey().toString(), keyValue.getValue().toString());
-            });
-        }
-        return map;
-    }
 
     /**
      * Generate kubernetes artifacts.
@@ -207,7 +182,6 @@ class KubernetesAnnotationProcessor {
         generateDeployment(deploymentModel, balxFilePath, outputDir);
         out.println();
         out.println("@kubernetes:Deployment \t\t\t - complete 1/1");
-        out.println();
 
         printKubernetesInstructions(outputDir);
     }
@@ -216,10 +190,10 @@ class KubernetesAnnotationProcessor {
     private void generateDeployment(DeploymentModel deploymentModel, String balxFilePath, String outputDir) throws
             KubernetesPluginException {
         String balxFileName = KubernetesUtils.extractBalxName(balxFilePath);
-        if (deploymentModel.getName() == null) {
+        if (isEmpty(deploymentModel.getName())) {
             deploymentModel.setName(getValidName(balxFileName) + DEPLOYMENT_POSTFIX);
         }
-        if (deploymentModel.getImage() == null) {
+        if (isEmpty(deploymentModel.getImage())) {
             deploymentModel.setImage(balxFileName + DOCKER_LATEST_TAG);
         }
         deploymentModel.addLabel(KubernetesConstants.KUBERNETES_SELECTOR_KEY, balxFileName);
@@ -413,7 +387,7 @@ class KubernetesAnnotationProcessor {
         int defaultReplicas = 1;
         deploymentModel.setReplicas(defaultReplicas);
         deploymentModel.addLabel(KubernetesConstants.KUBERNETES_SELECTOR_KEY, balxName);
-        deploymentModel.setEnv(getMap(null));
+        deploymentModel.setEnv(new HashMap<String, String>());
         deploymentModel.setImage(balxName + DOCKER_LATEST_TAG);
         deploymentModel.setBuildImage(true);
         deploymentModel.setPush(false);
@@ -421,206 +395,6 @@ class KubernetesAnnotationProcessor {
         return deploymentModel;
     }
 
-    /**
-     * Process annotations and create deployment model object.
-     *
-     * @param attachmentNode annotation attachment node.
-     * @return Deployment model object
-     */
-    DeploymentModel processDeployment(AnnotationAttachmentNode attachmentNode) throws KubernetesPluginException {
-        DeploymentModel deploymentModel = new DeploymentModel();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            DeploymentConfiguration deploymentConfiguration =
-                    DeploymentConfiguration.valueOf(keyValue.getKey().toString());
-            String annotationValue = resolveValue(keyValue.getValue().toString());
-            switch (deploymentConfiguration) {
-                case name:
-                    deploymentModel.setName(getValidName(annotationValue));
-                    break;
-                case labels:
-                    deploymentModel.setLabels(getMap(((BLangRecordLiteral) keyValue.valueExpr).keyValuePairs));
-                    break;
-                case enableLiveness:
-                    deploymentModel.setEnableLiveness(annotationValue);
-                    break;
-                case livenessPort:
-                    deploymentModel.setLivenessPort(Integer.parseInt(annotationValue));
-                    break;
-                case initialDelaySeconds:
-                    deploymentModel.setInitialDelaySeconds(Integer.parseInt(annotationValue));
-                    break;
-                case periodSeconds:
-                    deploymentModel.setPeriodSeconds(Integer.parseInt(annotationValue));
-                    break;
-                case username:
-                    deploymentModel.setUsername(annotationValue);
-                    break;
-                case env:
-                    deploymentModel.setEnv(getMap(((BLangRecordLiteral) keyValue.valueExpr).keyValuePairs));
-                    break;
-                case password:
-                    deploymentModel.setPassword(annotationValue);
-                    break;
-                case baseImage:
-                    deploymentModel.setBaseImage(annotationValue);
-                    break;
-                case push:
-                    deploymentModel.setPush(Boolean.valueOf(annotationValue));
-                    break;
-                case buildImage:
-                    deploymentModel.setBuildImage(Boolean.valueOf(annotationValue));
-                    break;
-                case image:
-                    deploymentModel.setImage(annotationValue);
-                    break;
-                case dockerHost:
-                    deploymentModel.setDockerHost(annotationValue);
-                    break;
-                case dockerCertPath:
-                    deploymentModel.setDockerCertPath(annotationValue);
-                    break;
-                case imagePullPolicy:
-                    deploymentModel.setImagePullPolicy(annotationValue);
-                    break;
-                case replicas:
-                    deploymentModel.setReplicas(Integer.parseInt(annotationValue));
-                    break;
-                default:
-                    break;
-            }
-        }
-        return deploymentModel;
-    }
-
-    /**
-     * Process annotations and create service model object.
-     *
-     * @param endpointName   ballerina service name
-     * @param attachmentNode annotation attachment node.
-     * @return Service model object
-     */
-    ServiceModel processServiceAnnotation(String endpointName, AnnotationAttachmentNode attachmentNode) throws
-            KubernetesPluginException {
-        ServiceModel serviceModel = new ServiceModel();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            ServiceConfiguration serviceConfiguration =
-                    ServiceConfiguration.valueOf(keyValue.getKey().toString());
-            String annotationValue = resolveValue(keyValue.getValue().toString());
-            switch (serviceConfiguration) {
-                case name:
-                    serviceModel.setName(getValidName(annotationValue));
-                    break;
-                case labels:
-                    serviceModel.setLabels(getMap(((BLangRecordLiteral) keyValue.valueExpr).keyValuePairs));
-                    break;
-                case serviceType:
-                    serviceModel.setServiceType(annotationValue);
-                    break;
-                case port:
-                    serviceModel.setPort(Integer.parseInt(annotationValue));
-                    break;
-                default:
-                    break;
-            }
-        }
-        if (serviceModel.getName() == null) {
-            serviceModel.setName(getValidName(endpointName) + SVC_POSTFIX);
-        }
-        return serviceModel;
-    }
-
-    /**
-     * Process annotations and create service model object.
-     *
-     * @param attachmentNode annotation attachment node.
-     * @return Service model object
-     */
-    PodAutoscalerModel processPodAutoscalerAnnotation(AnnotationAttachmentNode attachmentNode) throws
-            KubernetesPluginException {
-        PodAutoscalerModel podAutoscalerModel = new PodAutoscalerModel();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            PodAutoscalerConfiguration podAutoscalerConfiguration =
-                    PodAutoscalerConfiguration.valueOf(keyValue.getKey().toString());
-            String annotationValue = resolveValue(keyValue.getValue().toString());
-            switch (podAutoscalerConfiguration) {
-                case name:
-                    podAutoscalerModel.setName(getValidName(annotationValue));
-                    break;
-                case labels:
-                    podAutoscalerModel.setLabels(getMap(((BLangRecordLiteral) keyValue.valueExpr).keyValuePairs));
-                    break;
-                case cpuPercentage:
-                    podAutoscalerModel.setCpuPercentage(Integer.parseInt(annotationValue));
-                    break;
-                case minReplicas:
-                    podAutoscalerModel.setMinReplicas(Integer.parseInt(annotationValue));
-                    break;
-                case maxReplicas:
-                    podAutoscalerModel.setMaxReplicas(Integer.parseInt(annotationValue));
-                    break;
-                default:
-                    break;
-            }
-        }
-        return podAutoscalerModel;
-    }
-
-    /**
-     * Process annotations and create Ingress model object.
-     *
-     * @param serviceName    Ballerina service name
-     * @param attachmentNode annotation attachment node.
-     * @return Ingress model object
-     */
-    IngressModel processIngressAnnotation(String serviceName, AnnotationAttachmentNode attachmentNode) throws
-            KubernetesPluginException {
-        IngressModel ingressModel = new IngressModel();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            IngressConfiguration ingressConfiguration =
-                    IngressConfiguration.valueOf(keyValue.getKey().toString());
-            String annotationValue = resolveValue(keyValue.getValue().toString());
-            switch (ingressConfiguration) {
-                case name:
-                    ingressModel.setName(getValidName(annotationValue));
-                    break;
-                case labels:
-                    ingressModel.setLabels(getMap(((BLangRecordLiteral) keyValue.valueExpr).keyValuePairs));
-                    break;
-                case path:
-                    ingressModel.setPath(annotationValue);
-                    break;
-                case targetPath:
-                    ingressModel.setTargetPath(annotationValue);
-                    break;
-                case hostname:
-                    ingressModel.setHostname(annotationValue);
-                    break;
-                case ingressClass:
-                    ingressModel.setIngressClass(annotationValue);
-                    break;
-                case enableTLS:
-                    ingressModel.setEnableTLS(Boolean.parseBoolean(annotationValue));
-                    break;
-                default:
-                    break;
-            }
-        }
-        if (ingressModel.getName() == null || ingressModel.getName().length() == 0) {
-            ingressModel.setName(getValidName(serviceName) + INGRESS_POSTFIX);
-        }
-        if (ingressModel.getHostname() == null || ingressModel.getHostname().length() == 0) {
-            ingressModel.setHostname(getValidName(serviceName) + INGRESS_HOSTNAME_POSTFIX);
-        }
-        return ingressModel;
-    }
 
     /**
      * Extract key-store/trust-store file location from endpoint.
@@ -713,247 +487,9 @@ class KubernetesAnnotationProcessor {
         return null;
     }
 
-    /**
-     * Process Secrets annotations.
-     *
-     * @param attachmentNode Attachment Node
-     * @return List of @{@link SecretModel} objects
-     */
-    Set<SecretModel> processSecrets(AnnotationAttachmentNode attachmentNode) throws KubernetesPluginException {
-        Set<SecretModel> secrets = new HashSet<>();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            List<BLangExpression> secretAnnotation = ((BLangArrayLiteral) keyValue.valueExpr).exprs;
-            for (BLangExpression bLangExpression : secretAnnotation) {
-                SecretModel secretModel = new SecretModel();
-                List<BLangRecordLiteral.BLangRecordKeyValue> annotationValues =
-                        ((BLangRecordLiteral) bLangExpression).getKeyValuePairs();
-                for (BLangRecordLiteral.BLangRecordKeyValue annotation : annotationValues) {
-                    VolumeMountConfig volumeMountConfig =
-                            VolumeMountConfig.valueOf(annotation.getKey().toString());
-                    String annotationValue = resolveValue(annotation.getValue().toString());
-                    switch (volumeMountConfig) {
-                        case name:
-                            secretModel.setName(getValidName(annotationValue));
-                            break;
-                        case mountPath:
-                            secretModel.setMountPath(annotationValue);
-                            break;
-                        case data:
-                            List<BLangExpression> data = ((BLangArrayLiteral) annotation.valueExpr).exprs;
-                            secretModel.setData(getDataForSecret(data));
-                            break;
-                        case readOnly:
-                            secretModel.setReadOnly(Boolean.parseBoolean(annotationValue));
-                            break;
-                        default:
-                            break;
-                    }
-                }
-                secrets.add(secretModel);
-            }
-        }
-        return secrets;
-    }
-
-    /**
-     * Process ConfigMap annotations.
-     *
-     * @param attachmentNode Attachment Node
-     * @return Set of @{@link ConfigMapModel} objects
-     */
-    Set<ConfigMapModel> processConfigMap(AnnotationAttachmentNode attachmentNode) throws KubernetesPluginException {
-        Set<ConfigMapModel> configMapModels = new HashSet<>();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            List<BLangExpression> configAnnotation = ((BLangArrayLiteral) keyValue.valueExpr).exprs;
-            for (BLangExpression bLangExpression : configAnnotation) {
-                ConfigMapModel configMapModel = new ConfigMapModel();
-                List<BLangRecordLiteral.BLangRecordKeyValue> annotationValues =
-                        ((BLangRecordLiteral) bLangExpression).getKeyValuePairs();
-                for (BLangRecordLiteral.BLangRecordKeyValue annotation : annotationValues) {
-                    VolumeMountConfig volumeMountConfig =
-                            VolumeMountConfig.valueOf(annotation.getKey().toString());
-                    String annotationValue = resolveValue(annotation.getValue().toString());
-                    switch (volumeMountConfig) {
-                        case name:
-                            configMapModel.setName(getValidName(annotationValue));
-                            break;
-                        case mountPath:
-                            configMapModel.setMountPath(annotationValue);
-                            break;
-                        case isBallerinaConf:
-                            configMapModel.setBallerinaConf(Boolean.parseBoolean(annotationValue));
-                            break;
-                        case data:
-                            List<BLangExpression> data = ((BLangArrayLiteral) annotation.valueExpr).exprs;
-                            configMapModel.setData(getDataForConfigMap(data));
-                            break;
-                        case readOnly:
-                            configMapModel.setReadOnly(Boolean.parseBoolean(annotationValue));
-                            break;
-                        default:
-                            break;
-                    }
-                }
-                configMapModels.add(configMapModel);
-            }
-        }
-        return configMapModels;
-    }
-
-    /**
-     * Process PersistentVolumeClaim annotations.
-     *
-     * @param attachmentNode Attachment Node
-     * @return Set of @{@link ConfigMapModel} objects
-     */
-    Set<PersistentVolumeClaimModel> processPersistentVolumeClaim(AnnotationAttachmentNode attachmentNode) throws
-            KubernetesPluginException {
-        Set<PersistentVolumeClaimModel> volumeClaimModels = new HashSet<>();
-        List<BLangRecordLiteral.BLangRecordKeyValue> keyValues =
-                ((BLangRecordLiteral) ((BLangAnnotationAttachment) attachmentNode).expr).getKeyValuePairs();
-        for (BLangRecordLiteral.BLangRecordKeyValue keyValue : keyValues) {
-            List<BLangExpression> secretAnnotation = ((BLangArrayLiteral) keyValue.valueExpr).exprs;
-            for (BLangExpression bLangExpression : secretAnnotation) {
-                PersistentVolumeClaimModel claimModel = new PersistentVolumeClaimModel();
-                List<BLangRecordLiteral.BLangRecordKeyValue> annotationValues =
-                        ((BLangRecordLiteral) bLangExpression).getKeyValuePairs();
-                for (BLangRecordLiteral.BLangRecordKeyValue annotation : annotationValues) {
-                    VolumeClaimConfig volumeMountConfig =
-                            VolumeClaimConfig.valueOf(annotation.getKey().toString());
-                    String annotationValue = resolveValue(annotation.getValue().toString());
-                    switch (volumeMountConfig) {
-                        case name:
-                            claimModel.setName(getValidName(annotationValue));
-                            break;
-                        case mountPath:
-                            claimModel.setMountPath(annotationValue);
-                            break;
-                        case accessMode:
-                            claimModel.setAccessMode(annotationValue);
-                            break;
-                        case volumeClaimSize:
-                            claimModel.setVolumeClaimSize(annotationValue);
-                            break;
-                        case readOnly:
-                            claimModel.setReadOnly(Boolean.parseBoolean(annotationValue));
-                            break;
-                        default:
-                            break;
-                    }
-                }
-                volumeClaimModels.add(claimModel);
-            }
-        }
-        return volumeClaimModels;
-    }
-
-    private Map<String, String> getDataForSecret(List<BLangExpression> data) throws KubernetesPluginException {
-        Map<String, String> dataMap = new HashMap<>();
-        for (BLangExpression bLangExpression : data) {
-            Path dataFilePath = Paths.get(((BLangLiteral) bLangExpression).getValue().toString());
-            String key = String.valueOf(dataFilePath.getFileName());
-            String content = Base64.encodeBase64String(KubernetesUtils.readFileContent(dataFilePath));
-            dataMap.put(key, content);
-        }
-        return dataMap;
-    }
-
-    private Map<String, String> getDataForConfigMap(List<BLangExpression> data) throws KubernetesPluginException {
-        Map<String, String> dataMap = new HashMap<>();
-        for (BLangExpression bLangExpression : data) {
-            Path dataFilePath = Paths.get(((BLangLiteral) bLangExpression).getValue().toString());
-            String key = String.valueOf(dataFilePath.getFileName());
-            String content = new String(KubernetesUtils.readFileContent(dataFilePath), StandardCharsets.UTF_8);
-            dataMap.put(key, content);
-        }
-        return dataMap;
-    }
 
     private String getValidName(String name) {
         return name.toLowerCase(Locale.ENGLISH).replace("_", "-");
     }
 
-    /**
-     * Enum class for DeploymentConfiguration.
-     */
-    private enum DeploymentConfiguration {
-        name,
-        labels,
-        replicas,
-        enableLiveness,
-        livenessPort,
-        initialDelaySeconds,
-        periodSeconds,
-        imagePullPolicy,
-        namespace,
-        image,
-        env,
-        buildImage,
-        dockerHost,
-        username,
-        password,
-        baseImage,
-        push,
-        dockerCertPath
-    }
-
-    /**
-     * Enum class for svc configurations.
-     */
-    private enum ServiceConfiguration {
-        name,
-        labels,
-        serviceType,
-        port
-    }
-
-    /**
-     * Enum class for svc configurations.
-     */
-    private enum IngressConfiguration {
-        name,
-        labels,
-        hostname,
-        path,
-        targetPath,
-        ingressClass,
-        enableTLS
-    }
-
-    /**
-     * Enum class for pod autoscaler configurations.
-     */
-    private enum PodAutoscalerConfiguration {
-        name,
-        labels,
-        minReplicas,
-        maxReplicas,
-        cpuPercentage
-    }
-
-    /**
-     * Enum class for volume configurations.
-     */
-    private enum VolumeMountConfig {
-        name,
-        mountPath,
-        readOnly,
-        isBallerinaConf,
-        data
-    }
-
-    /**
-     * Enum class for volume configurations.
-     */
-    private enum VolumeClaimConfig {
-        name,
-        mountPath,
-        readOnly,
-        accessMode,
-        volumeClaimSize
-    }
 }
